+++ src/routes/+layout.svelte
@@ -1,5 +1,9 @@
 <script lang="ts">
 	import '../app.css';
+	import appleTouchIcon from '$lib/assets/apple-touch-icon.png';
+	import faviconIco from '$lib/assets/favicon.ico';
+	import favicon16 from '$lib/assets/favicon-16x16.png';
+	import favicon32 from '$lib/assets/favicon-32x32.png';
 	import favicon from '$lib/assets/favicon.svg';
 	import AppFooter from '$lib/components/AppFooter.svelte';
 	import Nav from '$lib/components/Nav.svelte';
@@ -45,7 +49,11 @@
 </script>
 
 <svelte:head>
-	<link rel="icon" href={favicon} />
+	<link rel="icon" type="image/png" sizes="32x32" href={favicon32} />
+	<link rel="icon" type="image/png" sizes="16x16" href={favicon16} />
+	<link rel="icon" href={faviconIco} sizes="any" />
+	<link rel="icon" href={favicon} type="image/svg+xml" />
+	<link rel="apple-touch-icon" href={appleTouchIcon} />
 	<title>{title}</title>
 </svelte:head>